Omaha #5308 - playlist purge will happen independently of message audio/metadata purge.

+++ edex/com.raytheon.bmh.dactransmit/src/com/raytheon/bmh/dactransmit/playlist/PlaylistMessageCache.java
@@ -123,6 +123,7 @@ import com.raytheon.uf.edex.bmh.msg.logging.ErrorActivity.BMH_COMPONENT;
  * Oct 06, 2015 4904       bkowal       Handle the case when the last purge time is not set.
  * Nov 04, 2015 5068       rjpeter      Switch audio units from dB to amplitude.
  * Feb 04, 2016 5308       bkowal       Utilize {@link DacPlaylistMessageMetadata}.
+ * Feb 15, 2016 5308       bkowal       Playlist file purge will happen independently of message purge.
  * </pre>
  * 
  * @author dgilling
@@ -869,22 +870,28 @@ public final class PlaylistMessageCache implements IAudioJobListener {
 
         @Override
         public Object call() {
+            DacPlaylistMessageId logMessageId = null;
             try {
                 for (DacPlaylistMessageId messageId : playlist.getMessages()) {
+                    logMessageId = messageId;
                     /*
-                     * The message metadata file will have already been removed
-                     * for this message if it has been updated. So, the rare
-                     * case in which message metadata does not exist on disk and
-                     * this is the last time the message has been referenced in
-                     * a playlist will be handled by the periodic purge task
-                     * because this message will not be present in active
-                     * messages.
+                     * Only allow the playlist purge to remove the message
+                     * information if this is the last version of the message (a
+                     * metadata file still remains) and the message has expired
+                     * (there are not any updated versions of the message
+                     * metadata file). All other cases will be handled by the
+                     * {@link PurgeTask}.
                      */
-                    if (isExpired(messageId)) {
+                    if (doesMessageFileExist(messageId) && isExpired(messageId)) {
                         purgeAudio(messageId);
                     }
                 }
+            } catch (Throwable e) {
+                logger.error("Error removing message audio for message: "
+                        + logMessageId + ".", e);
+            }
 
+            try {
                 /*
                  * Playlist may be deleted before the purge runs if a newer
                  * version of the playlist is received.
@@ -897,6 +904,7 @@ public final class PlaylistMessageCache implements IAudioJobListener {
                 logger.error("Error deleting playlist " + playlist.getPath()
                         + " from disk.", e);
             }
+
             return null;
         }
